% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azcast-color.R
\name{azcast-color}
\alias{azcast-color}
\alias{color_all}
\alias{color_show}
\alias{color_primary}
\alias{color_secondary}
\title{AZCAST brand colors}
\usage{
color_all() # show both primary and secondary colors

color_show(print = TRUE)

color_primary()

color_secondary()
}
\arguments{
\item{print}{whether to print the color vector}
}
\description{
AZCAST brand colors
}
\section{Primary palette}{
The basic blue and red. These are used most often.
}

\section{Secondary palette}{
These range from highlights to secondary colors to neutrals.
}

