% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azcast-package.R
\docType{package}
\name{azcast-package}
\alias{azcast}
\alias{azcast-package}
\title{azcast: Consolidated tools and templates for staff at University of Arizona's College of Applied Science and Technology}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Rmarkdown templates and other tools for use at the University of Arizona, Applied Computing program. It is a fork of numats/monash.
}
\author{
\strong{Maintainer}: Ryan Straight \email{ryanstraight@arizona.edu} (\href{https://orcid.org/0000-0002-6251-5662}{ORCID})

}
\keyword{internal}
