% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{letter}
\alias{letter}
\alias{exam}
\alias{workingpaper}
\alias{report}
\alias{memo}
\alias{assignment}
\title{Output formats for azcast documents}
\usage{
letter(...)

exam(...)

workingpaper(...)

report(...)

memo(...)

assignment(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[bookdown]{pdf_document2}} to
produce documents styled for AZCAST.
}
\author{
Ryan Straight (Original: Rob J Hyndman)
}
