% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo_get}
\alias{logo_get}
\title{Get azcast logo}
\source{
https://brand.arizona.edu/applying-the-brand/logo
}
\usage{
logo_get(
  path = ".",
  style = c("horiz", "stack", "block"),
  color = c("rgb", "black", "blue"),
  hq = FALSE,
  overwrite = TRUE,
  filename = NULL
)
}
\arguments{
\item{path}{the path to save the file to}

\item{style}{\code{horiz}, \code{stack} or \code{block}}

\item{color}{\code{rgb}, \code{black} or \code{blue}}

\item{hq}{TRUE for eps}

\item{overwrite}{TRUE for overwriting (should it enquire?), FALSE for not. Not implemented yet.}

\item{filename}{A new file name for the logo. Not implemented yet.}
}
\description{
A quick way of getting the University of Arizona logo.
}
