% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webheader.R
\name{logo_get_cast}
\alias{logo_get_cast}
\title{Get CAST webheader}
\source{
https://azcast.arizona.edu/
}
\usage{
logo_get_cast(path = ".", overwrite = TRUE, filename = NULL)
}
\arguments{
\item{path}{the path to save the file to}

\item{overwrite}{TRUE for overwriting (should it enquire?), FALSE for not. Not implemented yet.}

\item{filename}{A new file name for the logo. Not implemented yet.}
}
\description{
A quick way of getting the CAST webheader
}
