% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uarizona-package.R
\docType{package}
\name{uarizona-package}
\alias{uarizona}
\alias{uarizona-package}
\title{uarizona: Consolidated tools and templates for staff at uarizona University}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Rmarkdown templates and other tools for use at uarizona University, Department of Econometrics & Business Statistics.
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

Authors:
\itemize{
  \item Rob Hyndman \email{Rob.Hyndman@uarizona.edu}
}

Other contributors:
\itemize{
  \item Mitchell O'Hara-Wild (\href{https://orcid.org/0000-0001-6729-7695}{ORCID}) [contributor]
}

}
\keyword{internal}
