% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azcast-package.R
\docType{package}
\name{azcast-package}
\alias{azcast}
\alias{azcast-package}
\title{azcast: Templates for the College of Applied Science and Technology at the University of Arizona}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

RMarkdown templates and other tools for use in the College of Applied Science and Technology at the University of Arizona.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://osf.io/48zv2/}
}

}
\author{
\strong{Maintainer}: Ryan Straight \email{ryanstraight@arizona.edu} (\href{https://orcid.org/0000-0002-6251-5662}{ORCID})

}
\keyword{internal}
